backend/requirements.txt
fastapi
uvicorn
openai
openai-whisper
pydantic
python-multipart
python-dotenv